requirements.txt
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary==2.9.9
grpcio>=1.60
protobuf>=4.25
psycopg[binary]>=3.1
